added comment/new route (renamed /application/<application_id>/comments route
