+++ backend/src/main.rs
@@ -1,9 +1,9 @@
 #[macro_use]
 extern crate diesel;
 
-pub mod application;
 pub mod auth;
 pub mod campaigns;
+pub mod comment;
 pub mod cors;
 pub mod database;
 pub mod guard;
@@ -57,7 +57,7 @@ async fn main() {
             ],
         )
         .mount("/user", routes![user::get_user, user::get_user_campaigns])
-        .mount("/application", routes![application::create_comment])
+        .mount("/comment", routes![comment::create_comment])
         .launch()
         .await
         .unwrap();
